Update CMakeLists.txt
Issue #25 solved. 
Added if at line 45, based on this question:
https://stackoverflow.com/questions/2274006/cmake-invalid-numeric-argument-wextra

After that, a new error popped up concerning "m.lib" file not found.
Removed "m" at line 61, based on this question:
https://stackoverflow.com/questions/19333898/lnk1181-cannot-open-input-file-m-lib

Now the project can be built successfully and the example runs correctly.

PS: I'm not a computer scientist so I do not assure what I did is wrong, consider this pull request as a hint to solve Issue #25.